Move dynamic compiling to own class.
